Added 2 colors of Normal Block and the Hard Block sprites. Added acceleration of both ball and player bar. Improved collision with bricks and ball.

+++ Main/Main.java
@@ -78,7 +78,7 @@ public class Main
 	    		while(player.getScore() <= board.totalBlocks())
 	    		{
 	    			
-	    			ball.updatePos(ball.getHorzMovement(), ball.getVertMovement());  //Ball gets moved
+	    			ball.updatePos((int) ball.getHorzMovement(), (int) ball.getVertMovement());  //Ball gets moved
 	    			
 	    			char direction = ' ';
